fix(gitpod): possibly fix sync-await waits forever bug
We just call sync-done when the workspace is started for our web/API server to start up, through we
need to configure Redis and bunch of OAuth stuff, possibly to be told to the used in the future.

Also added the Rust Analyzer extension, and just an small tweak to don't distrub the devloper
when the Postgres server is being started by our bashrc hack.

Signed-off-by: Andrei Jiroh Eugenio Halili <andreijiroh@madebythepins.tk>