chore(global): add recommended settings for Dart Code extension, among other things
Also, set git.alwaysSignOff to match to the behavior in 'yarn commit:dco' and also added JSON
comments, through it shouldn't do that in the JSON spec. And by the way, I enabled
rust-analyzer.cargo.autoreload to automagically refresh project info via 'cargo metadata' when the
Cargo.toml file in the Hydralite CLI is changed.

Signed-off-by: Andrei Jiroh Eugenio Halili <andreijiroh@madebythepins.tk>