feat(commitizen): use an custom Commitizen adapter
Currently, the configuration is just the same, through some errors we're firing up if I call
directly through changing the path to the adapter script. Probably I just publish that to npmjs?
Also the Yarn lockfile was updated when I added cz-commitlint-config directory to the workspaces key
in the global package.json.

Signed-off-by: Andrei Jiroh Eugenio Halili <andreijiroh@madebythepins.tk>